+++ packages/schemas/src/v2/index.type.ts
@@ -41,6 +41,14 @@ export interface Settings {
    */
   maxFileSize?: number;
   /**
+   * The maximum amount of trigrams per document. Documents that exceed this maximum will not be indexed. Defaults to 20000
+   */
+  maxTrigramCount?: number;
+  /**
+   * Automatically delete stale repositories from the index. Defaults to true.
+   */
+  autoDeleteStaleRepos?: boolean;
+  /**
    * The interval (in milliseconds) at which the indexer should re-index all repositories. Repositories are always indexed when first added. Defaults to 1 hour (3600000 milliseconds).
    */
   reindexInterval?: number;