+++ packages/web/src/app/[domain]/layout.tsx
@@ -11,6 +11,7 @@ import { getSelectorsByUserAgent } from "react-device-detect";
 import { MobileUnsupportedSplashScreen } from "./components/mobileUnsupportedSplashScreen";
 import { MOBILE_UNSUPPORTED_SPLASH_SCREEN_DISMISSED_COOKIE_NAME } from "@/lib/constants";
 import { SyntaxReferenceGuide } from "./components/syntaxReferenceGuide";
+import { SyntaxGuideProvider } from "./components/syntaxGuideProvider";
 
 interface LayoutProps {
     children: React.ReactNode,
@@ -81,9 +82,9 @@ export default async function Layout({
         )
     }
     return (
-        <>
+        <SyntaxGuideProvider>
             {children}
             <SyntaxReferenceGuide />
-        </>
+        </SyntaxGuideProvider>
     )
 }