+++ packages/web/src/app/posthogProvider.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { NEXT_PUBLIC_POSTHOG_PAPIK, NEXT_PUBLIC_POSTHOG_UI_HOST, NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED } from '@/lib/environment.client'
+import { NEXT_PUBLIC_POSTHOG_PAPIK, NEXT_PUBLIC_POSTHOG_UI_HOST, NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED, NEXT_PUBLIC_PUBLIC_SEARCH_DEMO } from '@/lib/environment.client'
 import posthog from 'posthog-js'
 import { usePostHog } from 'posthog-js/react'
 import { PostHogProvider as PHProvider } from 'posthog-js/react'
@@ -45,12 +45,11 @@ export function PostHogProvider({ children }: { children: React.ReactNode }) {
             posthog.init(NEXT_PUBLIC_POSTHOG_PAPIK!, {
                 api_host: posthogHostPath,
                 ui_host: NEXT_PUBLIC_POSTHOG_UI_HOST,
-                capture_pageview: false, // Disable automatic pageview capture
+                person_profiles: 'identified_only',
+                capture_pageview: NEXT_PUBLIC_PUBLIC_SEARCH_DEMO, // @nocheckin Disable automatic pageview capture if we're not in public demo mode
                 autocapture: false, // Disable automatic event capture
                 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                /* @nocheckin HANDLE SELF HOSTED CASE
-                person_profiles: 'identified_only',
-                sanitize_properties: (properties: Record<string, any>, _event: string) => {
+                sanitize_properties: !NEXT_PUBLIC_PUBLIC_SEARCH_DEMO ? (properties: Record<string, any>, _event: string) => {
                     // https://posthog.com/docs/libraries/js#config
                     if (properties['$current_url']) {
                         properties['$current_url'] = null;
@@ -58,10 +57,9 @@ export function PostHogProvider({ children }: { children: React.ReactNode }) {
                     if (properties['$ip']) {
                         properties['$ip'] = null;
                     }
-    
+                
                     return properties;
-                }
-                */
+                } : undefined
             });
         } else {
             console.log("PostHog telemetry disabled");