+++ packages/web/src/lib/environment.ts
@@ -5,6 +5,7 @@ import { getEnv, getEnvBoolean, getEnvNumber } from "./utils";
 export const ZOEKT_WEBSERVER_URL = getEnv(process.env.ZOEKT_WEBSERVER_URL, "http://localhost:6070")!;
 export const SHARD_MAX_MATCH_COUNT = getEnvNumber(process.env.SHARD_MAX_MATCH_COUNT, 10000);
 export const TOTAL_MAX_MATCH_COUNT = getEnvNumber(process.env.TOTAL_MAX_MATCH_COUNT, 100000);
+export const SOURCEBOT_VERSION = getEnv(process.env.SOURCEBOT_VERSION, 'unknown')!;
 export const NODE_ENV = process.env.NODE_ENV;
 
 export const AUTH_SECRET = getEnv(process.env.AUTH_SECRET); // Generate using `npx auth secret`
@@ -25,3 +26,4 @@ export const SMTP_CONNECTION_URL = getEnv(process.env.SMTP_CONNECTION_URL);
 export const EMAIL_FROM = getEnv(process.env.EMAIL_FROM);
 
 export const SOURCEBOT_ROOT_DOMAIN = getEnv(process.env.SOURCEBOT_ROOT_DOMAIN, "localhost:3000")!;
+export const PUBLIC_SEARCH_DEMO = getEnvBoolean(process.env.PUBLIC_SEARCH_DEMO, false);